exception handling added for basic crud operations

+++ OnlineTutorFinderSystem/src/main/java/com/cg/project/controller/ParentController.java
@@ -14,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.cg.project.entity.Parent;
+import com.cg.project.exceptions.GlobalException;
 import com.cg.project.service.ParentServiceImpl;
 
 @RestController
@@ -25,21 +26,44 @@ public class ParentController {
 
 	@GetMapping("/getParentById/{pid}")
 	public Optional<Parent> getParentById(@PathVariable("pid") int parentId) {
-		return service.getParentById(parentId);
+		Optional<Parent> test = service.getParentById(parentId);
+		if (test.isEmpty()) {
+			throw new GlobalException("id is not present");
+		}
+
+		return test;
 	}
 
 	@PostMapping("/addParent")
 	public Parent addParent(@RequestBody Parent parent) {
-		return service.addParent(parent);
+		Parent test = service.addParent(parent);
+		if (test.getParentName() == null || test.getParentPhone() == null || test.getParentEmail() == null
+				|| test.getParentPhone().length() != 10) {
+			throw new GlobalException("Enter values for body correctly or check the phone no");
+		}
+		return test;
 	}
 
 	@PutMapping("/updateParent")
 	public Parent updateParent(@RequestBody Parent parent) {
-		return service.updateParent(parent);
+		Parent test = service.updateParent(parent);
+		if (test.getParentId() == 0 || test.getParentName() == null || test.getParentPhone() == null
+				|| test.getParentEmail() == null) {
+			throw new GlobalException("Id not present or enter the body correctly.");
+		}
+		Optional<Parent> test1 = service.getParentById(test.getParentId());
+		if (test1.isEmpty()) {
+			throw new GlobalException("id is not present for updating");
+		}
+		return test;
 	}
 
 	@DeleteMapping("/deleteParent/{pid}")
 	public void deleteParent(@PathVariable("pid") int parentId) {
+		Optional<Parent> test = service.getParentById(parentId);
+		if (test.isEmpty()) {
+			throw new GlobalException("id is not present for deleting");
+		}
 		service.deleteParent(parentId);
 	}
